added styling

+++ client/src/components/shoppingCart.js
@@ -30,39 +30,6 @@ class ShoppingCart extends Component {
       .catch(err => console.log(err))
   }
 
-
-  orderList = () => {
-    if (this.state.orders.length < 1) {
-      return(
-        <div>
-          <p>Basket is empty</p>
-        </div>
-      )
-    } else {
-      return this.state.orders.map(order => {
-        if (order.milkType === 'n/a') {
-          return(
-            <div key={order.id}>
-              <p>{order.name}</p>
-              <p>£{Number(order.price).toFixed(2)}</p>
-              <img src={order.imgUrl} alt={order.name} />
-              <a href="#" onClick={() => {this.deleteProduct(order.id)}}>delete</a>
-            </div>
-          )
-        } else {
-          return(
-            <div key={order.id}>
-              <p>{order.name}, {order.milkType}, {order.size}</p>
-              <p>£{Number(order.price).toFixed(2)}</p>
-              <img src={order.imgUrl} alt={order.name} />
-              <a href="#" onClick={() => {this.deleteProduct(order.id)}}>delete</a>
-            </div>
-          )
-        }
-      })
-    }
-  }
-
   deleteProduct = (id) => {
     axios.delete(`/carts/${id}`)
       .then(res => console.log(res.data))
@@ -74,31 +41,35 @@ class ShoppingCart extends Component {
   }
 
   onSubmit = (e) => {
-    e.preventDefault();
-
-    const submitOrder = {
-      order: this.state.orders
-    };
-
-    
-    axios.post('/orders', submitOrder)
-      .then(res => console.log(res.data))
-      .catch(err => console.log("Error: " + err));
-
-    axios.delete('/carts')
-      .then(res => console.log(res.data))
-      .catch(err => console.log("Error: " + err));
-    
-    this.setState({
-      orders: [],
-      isSubmit: true
-    })
+    if(this.state.order.length > 0) {
+      e.preventDefault();
+  
+      const submitOrder = {
+        order: this.state.orders
+      };
+  
+      
+      axios.post('/orders', submitOrder)
+        .then(res => console.log(res.data))
+        .catch(err => console.log("Error: " + err));
+  
+      axios.delete('/carts')
+        .then(res => console.log(res.data))
+        .catch(err => console.log("Error: " + err));
+      
+      this.setState({
+        orders: [],
+        isSubmit: true
+      })
+    } else {
+      return null;
+    }
   }
 
   confirmSubmit = () => {
     if (this.state.isSubmit === true) {
       return (
-        <div>
+        <div className="confirm-pop">
           <p>Your order has been sent</p>
         </div>
       )
@@ -107,27 +78,89 @@ class ShoppingCart extends Component {
 
   total = () => {
     let total = 0
-
     this.state.orders.map(order => {
       total = total + order.price
     });
-
     return (
-      <div>
-        <p>Total: £{Number(total).toFixed(2)}</p>
+      <div className="right">
+        <p className="cart-total">Total: £{Number(total).toFixed(2)}</p>
       </div>
     )
+  }
+
+  orderList = () => {
+    if (this.state.orders.length < 1) {
+      return(
+        <tr>
+          <td>
+            <p>Basket is empty</p>
+          </td>
+        </tr>
+      )
+    } else {
+      return this.state.orders.map(order => {
+        if (order.milkType === 'n/a') {
+          return(
+            <tr class="border-bottom" key={order.id}>
+              <td width="30%">
+                <img className="small-img" src={order.imgUrl} alt={order.name} />
+              </td>
+              <td width="20%">
+                <p>{order.name}</p>
+                <a href="#" onClick={() => {this.deleteProduct(order.id)}}>remove</a>
+              </td>
+              <td width="50%">
+                <p className="right">£{Number(order.price).toFixed(2)}</p>
+              </td>
+            </tr>
+          )
+        } else {
+          return(
+            <tr class="border-bottom" key={order.id}>
+              <td width="30%">
+                <img className="small-img" src={order.imgUrl} alt={order.name} />
+              </td>
+              <td width="20%">
+                <p>{order.name}</p>
+                <p>{order.milkType} | {order.size}</p>
+                <a href="#" onClick={() => {this.deleteProduct(order.id)}}>remove</a>
+              </td>
+              <td width="50%">
+                <p className="right">£{Number(order.price).toFixed(2)}</p>
+              </td>
+            </tr>
+          )
+        }
+      })
+    }
   }
 
   render() {
     return (
       <div className="container">
-        Shopping Cart
-        {this.orderList()}
-        {this.total()}
-        <form onSubmit={this.onSubmit}>
-          <input type="submit" value="Submit your order" />
-        </form>
+        <h1 className="heading">Shopping Cart</h1>
+        <table className="table">
+          <tbody>
+            {this.orderList()}
+            <tr>
+              <td width="30%"></td>
+              <td width="20%"></td>
+              <td width="60%">{this.total()}</td>
+            </tr>
+            <tr>
+              <td width="30%"></td>
+              <td width="20%"></td>
+              <td width="60%">
+                <form onSubmit={this.onSubmit}>
+                  <div className="right">
+                    <input className="submit-btn" type="submit" value="Submit your order" />
+                  </div>
+                </form>
+              </td>
+            </tr>
+          </tbody>
+        </table>
+      
         {this.confirmSubmit()}
       </div>
     )
